requirements.txt
torch
torchvision
torchaudio
numpy
scipy
scikit-image
opencv-python
pillow
pandas
pyyaml
omegaconf
hydra-core
tqdm
matplotlib
seaborn
tensorboard
einops
albumentations
medpy
pytest
black
ruff
mypy